pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "rsspot"
version = "0.1.0"
description = "Async-first Python SDK and CLI for Rackspace Spot"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Halceon" }
]
dependencies = [
  "httpx>=0.28.1",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.8.0",
  "pyjwt>=2.10.1",
  "pyyaml>=6.0.2",
  "tomli-w>=1.2.0",
  "typer>=0.15.2",
  "rich>=13.9.4",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "pytest-asyncio>=0.25.3",
  "ruff>=0.11.0",
  "mypy>=1.15.0",
  "types-pyyaml>=6.0.12.20250915",
]

[project.scripts]
rsspot = "rsspot.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/rsspot"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "ASYNC"]

[tool.mypy]
python_version = "3.11"
strict = true
pretty = true
warn_unused_configs = true
mypy_path = "src"

[dependency-groups]
dev = [
  "pytest>=8.3.5",
  "pytest-asyncio>=0.25.3",
  "ruff>=0.11.0",
  "mypy>=1.15.0",
  "types-pyyaml>=6.0.12.20250915",
]
